deployment/Leo/ordering-helper-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
python-dotenv==1.0.0
PyMySQL==1.1.0
line-bot-sdk==3.5.0
google-genai==1.28.0
azure-cognitiveservices-speech==1.34.0
Pillow==10.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0

# Google Cloud Storage (用於語音檔上傳)
google-cloud-storage==2.10.0

# Google Cloud Translation API (用於多語言支援)
google-cloud-translate==3.12.0